backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
celery==5.4.0
elasticsearch==8.15.1
pydantic==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
